+++ src/components/game/winner-component/WinnerComponent.jsx
