+++ src/pages/offline-game/PageOfflineGame.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react'
 import AppTitle from '../../components/app-title/AppTitle'
 import MatchHeaderArea from '../../components/match-header-area/MatchHeaderArea'
-import OfflineMatchArea from '../../components/offline-match-area/OfflineMatchArea'
-import TurnIndicator from '../../components/turn-indicator/TurnIndicator'
-import WinnerComponent from '../../components/winner-component/WinnerComponent'
+import OfflineMatchArea from '../../components/game/offline-match-area/OfflineMatchArea'
+import TurnIndicator from '../../components/game/turn-indicator/TurnIndicator'
+import WinnerComponent from '../../components/game/winner-component/WinnerComponent'
 import ForfeitMatchFooter from '../../components/footers/forfeit-match-footer/ForfeitMatchFooter'
 
 function PageOfflineGame () {
